+++ RandomBot/Modules/NicknameModule/BotNicknameModule.cs
@@ -1,5 +1,4 @@
 ﻿using Discord.Commands;
-using Discord.WebSocket;
 using System.Threading.Tasks;
 
 namespace RandomBot.Modules.NicknameModule